@@ -1,4 +1,4 @@
-package android.support.v4.media;
+package androidx.media;
 
 import android.media.AudioAttributes;
 import android.media.AudioFocusRequest;
@@ -7,11 +7,12 @@ import android.media.AudioManager.OnAudioFocusChangeListener;
 import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.IntDef;
-import android.support.annotation.NonNull;
-import android.support.annotation.RequiresApi;
 
 import java.lang.annotation.Retention;
+
+import androidx.annotation.IntDef;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 
 import static java.lang.annotation.RetentionPolicy.SOURCE;
 
